refurbished encoders, added vims events for recording and parameter binding, some cleanup and bugfixes.
git-svn-id: svn://code.dyne.org/veejay/trunk@586 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-ng/bio2jack/bio2jack.c
@@ -577,8 +577,8 @@ JACK_callback(nframes_t nframes, void *arg)
          (&drv->callback_buffer2, &drv->callback_buffer2_size,
           jackBytesAvailable))
       {
-        ERR("allocated %lu bytes, need %lu bytes\n",
-            drv->callback_buffer2_size, jackBytesAvailable);
+        ERR("allocated %lu bytes, need %d bytes\n",
+            drv->callback_buffer2_size, (int)jackBytesAvailable);
         return -1;
       }
